refactor: summarize.tsの責務を分割した
