+++ apps/functions/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@google/genai": "^1.6.0",
     "cors": "^2.8.5",
+    "dayjs": "^1.11.13",
     "express": "^4.21.2",
     "express-promise-router": "^4.1.1",
     "express-validator": "^7.2.1",
